Initial commit
This is a simple utility that dumps regions of memory-mapped registers
for two implementations and then compares them register-by-register.

It is a situational, unpolished tool that i hacked quickly to debug my
own projects. There is likely a million edge cases that are not handled
and it may require some modification to work for boards that are a
little more tricky to debug as a remote GDB target.

Signed-off-by: Duszku <duszku511@gmail.com>